kvserver/rangefeed: move perRangeEventSink to rangefeed pacakge
Previously, perRangeEventSink was defined in pkg/server/node.go. This patch
relocates it to the rangefeed package to facilitate future commits testing
within the rangefeed package itself.

Part of: #126560
Release note: none
